Fix dropdown position on modals
SDCP-264

When the viewport is small, a dropdown inside a modal tends to drop up,
and it can overflow the modal body (which hides part of the dropdown).
After this PR the dropdown will drop down because even if some part gets
hidden you can always scroll down too see it.